feat: allow downloading verifier code from OpenVM assets
